Fix handling todo state of repeat task whose timestamp has
no "hh:mm" part.
The todo state of repeat task is derived by that the time of its
scheduled/deadline timestamp is earlier than the day start of the next
day.
When the timestamp has no "hh:mm" part, the hour and minute part
of its encoded time is set to zero.
But it is not correct to compare with the next day's day start.
So adjust the encoded time to the day start time.