Prevent moving cursor to the end of line when the line does not fit
the window.

`next-line' moves the cursor to the end of buffer when the cursor is
already at the last line and signals `end-of-buffer'.
But when the last line does not fit the current window's width,
`next-line' does not signal `end-of-buffer'.
So check that the cursor is already at the last line to prevent
moving the cursor to the end of buffer.